refactor(reborn): split capability host module
